# [Improvement] Optimize Gateway Artifact Generation with Asynchronous Processing