## Description

Refactored `SynapseArtifactGenerator.generateGatewayArtifact()` to process API runtime artifacts asynchronously, significantly improving performance for bulk operations while maintaining thread-safety and proper error handling.

## Related Issue
Fix: https://github.com/wso2/api-manager/issues/4012

## Approach
Implemented parallel processing using `CompletableFuture` while preserving the original method signature and behavior:

## Performance Improvement Summary

Carried out 10 APIM clustered setup startups, first half without the improvement and the rest with the async improvement.

- Total APIs deployed per run: 1,500

- **Time Reduction:** 1m 4.5s (37.7% faster)
- **Throughput Increase:** 5.32 APIs/second (60.5% improvement)
- **Most Consistent:** Async implementation shows less variance between runs

<img width="899" height="555" alt="Image" src="https://github.com/user-attachments/assets/9443a41a-9218-4b76-b487-ca702617edbf" /> <br>

| Metric | Sync Implementation | Async Implementation | Improvement |
|--------|-------------------|---------------------|-------------|
| **Average Runtime** | 2m 51s (171.03s) | 1m 46.5s (106.47s) | **37.7% faster** |
| **Average Throughput** | 8.80 APIs/sec | 14.12 APIs/sec | **60.5% higher** |